:sparkles: implementing employee extend user

+++ petshelter/src/main/java/com/ironhack/petshelter/controller/EmployeeController.java
@@ -1,10 +1,7 @@
 package com.ironhack.petshelter.controller;
 
 import com.ironhack.petshelter.dto.EmployeeDTO;
-import com.ironhack.petshelter.dto.VeterinarianDTO;
-import com.ironhack.petshelter.model.Adopter;
 import com.ironhack.petshelter.model.Employee;
-import com.ironhack.petshelter.model.Veterinarian;
 import com.ironhack.petshelter.service.EmployeeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -37,12 +34,12 @@ public class EmployeeController {
      */
     @GetMapping("/employees/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public Employee getEmployeeById(@PathVariable Integer id) {
+    public Employee getEmployeeById(@PathVariable Long id) {
         return employeeService.getEmployeeById(id);
     }
 
     @PatchMapping("/employees/{id}")
-    public Employee update(@PathVariable Integer id, @RequestBody EmployeeDTO employeeDTO) {
+    public Employee update(@PathVariable Long id, @RequestBody EmployeeDTO employeeDTO) {
         return employeeService.update(id, employeeDTO);
     }
